15894 int -> unsigned long

+++ 15894/main.cpp
@@ -2,7 +2,7 @@
 
 int main()
 {
-    int n;
+    unsigned long  n;
 
     std::cin >> n;
